+++ src/test/java/roomescape/fake/FakeReservationDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.fake;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
+import roomescape.dao.ReservationDao;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
