+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -10,9 +10,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.common.exception.DuplicatedException;
-import roomescape.dao.FakeReservationDao;
-import roomescape.dao.FakeReservationTimeDao;
-import roomescape.dao.FakeThemeDao;
+import roomescape.fake.FakeReservationDao;
+import roomescape.fake.FakeReservationTimeDao;
+import roomescape.fake.FakeThemeDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dto.request.ReservationRequestDto;
 import roomescape.dto.response.ReservationResponseDto;
